+++ ChangeLog.md
@@ -11,6 +11,7 @@ TODO : pour les options commencer à voir si possible de crééer des tab par el
 
 # RELEASE 1.36
 
+- FIX : Security issue with ticket *1.36.1* - **22/02/2022**
 - NEW : Factoring and deprecate Context::getRootUrl() see getControllerUrl()  *1.36.0* - **09/02/2023**
 - NEW : jSignature lib included in externalaccess *1.35.0* - **17/11/2022**